pyproject.toml
[tool.poetry]
name = "bilibili_storytree"
version = "0.1.1"
description = "API to create story graph for interactivate video in bilibili"
authors = ["Phoenix Grey <phoenix.grey0108@gmail.com>"]
readme = "README.md"
repository = "https://github.com/phigrey/bilibili-storytree.git"

[tool.poetry.dependencies]
python = "^3.8"
PyExecJS = "^1.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
